requirements.txt
amqp==5.2.0
asgiref==3.7.2
async-timeout==4.0.3
billiard==4.2.1
celery==5.4.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
Django==3.2
django-ckeditor==6.7.1
django-cors-headers==4.3.1
django-js-asset==2.2.0
django-redis==5.4.0
django-richtexteditor==0.0.1.dev3
djangorestframework==3.14.0
idna==3.10
kombu==5.4.2
numpy==2.0.2
opencv-python==4.10.0.84
pillow==10.3.0
prompt_toolkit==3.0.48
python-dateutil==2.9.0.post0
pytz==2023.3.post1
redis==5.1.0
requests==2.32.3
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.12.2
tzdata==2023.3
urllib3==2.2.3
vine==5.1.0
wcwidth==0.2.13
